Move things around some more
