+++ index.html
@@ -4,8 +4,6 @@
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <title>Gridcannon</title>
-  <script language="javascript" src="./libs/lz-string.min.js"></script>
-  <link rel="stylesheet" href="styles.css" type="text/css">
 </head>
 <body>
   <section class="nosupport" hidden>
