+++ src/app/game.consts.js
